guest-process/requirements.txt
google-cloud-secret-manager==1.0.0
google-cloud-bigquery
google-cloud-storage
python-dateutil
requests
google-api-core